study-plan-ml-system/requirements.txt
Flask==2.0.1
scikit-learn==0.24.2
pandas==1.2.4
numpy==1.20.3
xgboost==1.4.2
lightgbm==3.2.1
tensorflow==2.5.0
matplotlib==3.4.2
seaborn==0.11.1
jupyter==1.0.0
pytest==6.2.4
joblib==1.0.1
pydantic==1.8.2
python-dotenv==0.19.0
mysql-connector-python==8.0.26